new proxy extension
