+++ edc-extensions/dataplane/dataplane-proxy/dataplane-public-api-v2/src/main/java/org/eclipse/tractusx/edc/dataplane/proxy/api/DataPlanePublicApiV2Extension.java
@@ -19,13 +19,10 @@
 
 package org.eclipse.tractusx.edc.dataplane.proxy.api;
 
-import org.eclipse.edc.connector.dataplane.http.params.HttpRequestFactory;
-import org.eclipse.edc.connector.dataplane.http.spi.HttpRequestParamsProvider;
 import org.eclipse.edc.connector.dataplane.spi.Endpoint;
 import org.eclipse.edc.connector.dataplane.spi.iam.DataPlaneAuthorizationService;
 import org.eclipse.edc.connector.dataplane.spi.iam.PublicEndpointGeneratorService;
 import org.eclipse.edc.connector.dataplane.spi.pipeline.PipelineService;
-import org.eclipse.edc.http.spi.EdcHttpClient;
 import org.eclipse.edc.runtime.metamodel.annotation.Configuration;
 import org.eclipse.edc.runtime.metamodel.annotation.Extension;
 import org.eclipse.edc.runtime.metamodel.annotation.Inject;
@@ -39,7 +36,6 @@ import org.eclipse.edc.web.spi.WebService;
 import org.eclipse.edc.web.spi.configuration.ApiContext;
 import org.eclipse.edc.web.spi.configuration.PortMapping;
 import org.eclipse.edc.web.spi.configuration.PortMappingRegistry;
-import org.eclipse.tractusx.edc.dataplane.http.pipeline.ProxyHttpDataSourceFactory;
 import org.eclipse.tractusx.edc.dataplane.proxy.api.controller.DataPlanePublicApiV2Controller;
 
 import java.util.concurrent.Executors;
@@ -77,10 +73,6 @@ public class DataPlanePublicApiV2Extension implements ServiceExtension {
     private PublicEndpointGeneratorService generatorService;
     @Inject
     private Hostname hostname;
-    @Inject
-    private EdcHttpClient httpClient;
-    @Inject
-    private HttpRequestParamsProvider paramsProvider;
 
     @Override
     public String name() {
@@ -112,16 +104,9 @@ public class DataPlanePublicApiV2Extension implements ServiceExtension {
         if (publicApiResponseUrl != null) {
             generatorService.addGeneratorFunction("ProxyHttpData", () -> Endpoint.url(publicApiResponseUrl));
         }
-
+        
         var publicApiController = new DataPlanePublicApiV2Controller(pipelineService, executorService, authorizationService);
         webService.registerResource(ApiContext.PUBLIC, publicApiController);
-
-        var monitor = context.getMonitor();
-
-        var httpRequestFactory = new HttpRequestFactory();
-
-        var sourceFactory = new ProxyHttpDataSourceFactory(httpClient, paramsProvider, monitor, httpRequestFactory);
-        pipelineService.registerFactory(sourceFactory);
     }
 
     @Settings
